% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv_VCM}
\alias{cv_VCM}
\title{Cross-validation to select bandwidth for VCM}
\usage{
cv_VCM(data, h_seq, k = 5, d = 1)
}
\arguments{
\item{data}{a list containing matrices, each of which represents a sample of the
data in the VCM model. Each matrix should have a time column as the first column,
followed by p columns of covariates, and a final column of response values.}

\item{h_seq}{a sequence of candidate bandwidths to be evaluated in the
cross-validation.}

\item{k}{the number of folds in the cross-validation.}

\item{d}{the degree of the polynomial trend to be fitted in the VCM estimation.}
}
\value{
The value of \code{h} that minimizes the cross-validation mean squared error.
}
\description{
This function uses cross-validation to select the bandwidth parameter \code{h}
for the VCM estimation method.
}
\examples{
h_seq <- seq(0.1,1,length=10)
cv_VCM(VCMdata,h_seq,k=5,d=1)

}
