% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localp_Mean.R
\name{est_Mean}
\alias{est_Mean}
\title{Estimate the mean function at give time points}
\usage{
est_Mean(data, t_points, h = 0.1, d = 1)
}
\arguments{
\item{data}{A list of matrices, where each matrix contains the data for a single subject.}

\item{t_points}{The time points at which the mean function are estimated.}

\item{h}{The bandwidth parameter for the kernel function.}

\item{d}{The degree of the polynomial used in the local polynomial regression.}
}
\value{
A matrix representing the estimated mean function at the given time points.
}
\description{
This function estimates the mean function at given time poitns using local polynomial regression estimator with an Epanechnikov kernel.
}
\examples{
data <- rp_generate(n=10, m=20, mean_list=list(function(x) (x+1)^2, function(x) (x-1)^2),
            cov_list=list(function(x,y) exp(-abs(x-y)), function(x,y) exp(-2*abs(x-y))),
            sig=0.1)
est_Mean(data, t_points = seq(0,1,0.1))

}
