% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localp_VCM.R
\name{est_VCM}
\alias{est_VCM}
\title{Estimate the coefficients of a VCM model}
\usage{
est_VCM(data, t_points, h = 0.1, d = 1)
}
\arguments{
\item{data}{A list of data frames, each of which represents the observed data for one element.}

\item{t_points}{A vector of time points at which to estimate the VCM coefficients.}

\item{h}{The bandwidth for the Epanechnikov kernel. Default is 0.1.}

\item{d}{The degree of the polynomial for the local linear regression. Default is 1.}
}
\value{
A matrix of estimated coefficients, with each row representing the estimated coefficients for one element.
}
\description{
Given a list of data, estimate the coefficients of a VCM model at given time points
}
\examples{
data_list <- rp_VCM_generate(n=3,m=10,coef_list=list(f1,f2),mean_list=list(mean1,mean2),
                    cov_list=list(cov1,cov2), sig=0.1)
t_points <- seq(0, 1, by = 0.1)
est <- est_VCM(data_list, t_points)
est

}
