% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localp_Mean.R
\name{Mean_inference}
\alias{Mean_inference}
\title{Mean Inference for Local Polynomial Regression}
\usage{
Mean_inference(data, bstime = 1000, h = NULL, t_points, d = 1, alpha = 0.05)
}
\arguments{
\item{data}{A list of data frames, where each data frame contains the time series data for one subject. The first column of each data frame should contain the time points and the remaining columns should contain the observations at each time point.}

\item{bstime}{The number of bootstrap replicates.}

\item{h}{The bandwidth parameter for the local polynomial regression. If not provided, the bandwidth will be selected using the biweight cross-validation method.}

\item{t_points}{The time points at which to estimate the mean function.}

\item{d}{The degree of the local polynomial regression.}

\item{alpha}{The level of the confidence interval.}
}
\value{
A list containing the following components:
\item{muhat}{An estimate of the mean function at the specified time points.}
\item{low}{The lower bound of the confidence interval.}
\item{up}{The upper bound of the confidence interval.}
\item{band}{The average width of the confidence interval.}
\item{sd}{The standard deviation of the estimated mean function.}
}
\description{
This function performs mean inference for local polynomial regression.
}
\examples{
data <- rp_generate(n=10, m=20, mean_list=list(function(x) (x+1)^2, function(x) (x-1)^2),
            cov_list=list(function(x,y) exp(-abs(x-y)), function(x,y) exp(-2*abs(x-y))),
            sig=0.1)

Mean_inference(data = data, t_points = seq(0,1,0.1))

}
