% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprocess.R
\name{rp_generate}
\alias{rp_generate}
\title{Generate random functional data}
\usage{
rp_generate(
  n,
  m,
  mean_list,
  cov_list,
  dis_vector = NULL,
  covmatrix = NULL,
  sig = 0.1,
  depend = F,
  domain = c(0, 1)
)
}
\arguments{
\item{n}{The number of elements to generate}

\item{m}{The number of time points for each element}

\item{mean_list}{A list of length p containing mean functions for each of the p processes}

\item{cov_list}{A list of length p containing covariance functions for each of the p processes}

\item{dis_vector}{A vector of length p specifying the distribution to use for each of the p processes. If NULL,
'normal' distribution will be used for all processes.}

\item{covmatrix}{A p x p covariance matrix to be used when generating correlated processes. If NULL, an
identity matrix will be used.}

\item{sig}{The standard deviation of the error to add to the generated functional data}

\item{depend}{A boolean indicating whether to generate data with dependent errors or not}

\item{domain}{A vector of length 2 specifying the domain of the time points}
}
\value{
A list of length n, where each element is a matrix with the generated functional data. The matrix
  has nrow = m_i (the number of time points for the element) and ncol = p + 1 (the number of processes plus
  time points). The first column contains the time points and the remaining columns contain the functional
  observations.
}
\description{
This function generates random functional data with error. The data is generated by simulating a set of functional
observations for each of the n elements, which are assumed to have the same number of time points m. The function
generates a list of length n with each element of the list corresponding to a set of functional observations.
}
\examples{
# Generate random functional data with 10 elements, 20 time points for each element, 2 processes,
# mean functions (x+1)^2 and (x-1)^2, and covariance functions exp(-|x-y|) and exp(-2*|x-y|).
# The data will have independent errors with standard deviation 0.1.
rp_generate(n=10, m=20, mean_list=list(function(x) (x+1)^2, function(x) (x-1)^2),
            cov_list=list(function(x,y) exp(-abs(x-y)), function(x,y) exp(-2*abs(x-y))),
            sig=0.1)

}
