% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprocess.R
\name{rp_VCM_generate}
\alias{rp_VCM_generate}
\title{Generate data with a varying coefficient model}
\usage{
rp_VCM_generate(
  n,
  m,
  coef_list,
  mean_list,
  cov_list,
  dis_vector = NULL,
  covmatrix = NULL,
  sig = 0.1,
  depend = F,
  domain = c(0, 1)
)
}
\arguments{
\item{n}{Number of subjects.}

\item{m}{Number of time points for each subject.}

\item{coef_list}{A list of functions defining the varying coefficients for each covariate.}

\item{mean_list}{A list of means for the multivariate normal distribution used to generate the random effects.}

\item{cov_list}{A list of covariance matrices for the multivariate normal distribution used to generate the random effects.}

\item{dis_vector}{A vector of covariate values for each subject, defaults to NULL.}

\item{covmatrix}{A covariance matrix for the covariates, defaults to NULL.}

\item{sig}{Standard deviation of the error term, defaults to 0.1.}

\item{depend}{A logical indicating whether the error term is dependent on the outcome, defaults to FALSE.}

\item{domain}{A vector specifying the domain of the time points, defaults to c(0,1).}
}
\value{
A list of matrices representing the generated longitudinal data.
Each matrix has m columns, where the first column represents time and the remaining columns represent covariates and outcomes.
Each row represents an observation for a single subject at a single time point.
}
\description{
This function generates simulated longitudinal data with a varying coefficient model.
}
\examples{
n <- 50
m <- 10
coef_list <- list(function(x){x^2}, function(x){sin(pi*x)}, function(x){cos(pi*x)})
mean_list <- c(0,0,0)
cov_list <- list(diag(3),diag(3),diag(3))
dis_vector <- runif(n*3,0,1)
rp_VCM_generate(n=n, m=m, coef_list=coef_list, mean_list=mean_list, cov_list=cov_list, dis_vector=dis_vector)

}
