% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localp_VCM.R
\name{VCM_inference}
\alias{VCM_inference}
\title{Estimate and infer variance component model}
\usage{
VCM_inference(data, bstime = 1000, h = NULL, t_points, d = 1, alpha = 0.05)
}
\arguments{
\item{data}{A list of data frames, each containing data for one cluster.}

\item{bstime}{Number of bootstrap samples for confidence interval estimation.}

\item{h}{The bandwidth for the local polynomial regression. If NULL, the bandwidth will be chosen using cross-validation.}

\item{t_points}{A vector of time points to evaluate the estimate.}

\item{d}{The degree of the local polynomial regression.}

\item{alpha}{The significance level for the confidence interval.}
}
\value{
A list containing the estimate of the coefficients (betahat), the lower and upper bounds of the confidence interval (low and up), the average bandwidth (band), and the estimated standard deviation (sd).
\item{betahat}{An estimate of the coefficient functions at the specified time points.}
\item{low}{The lower bound of the confidence band.}
\item{up}{The upper bound of the confidence band.}
\item{band}{The average width of the confidence band.}
\item{sd}{The standard deviation of the estimated function.}
}
\description{
This function estimates the variance component model using local polynomial regression and performs inference for the coefficients.
}
\examples{
data(VCMdata)
VCM_inference(data = VCMdata, bstime = 100, h = NULL, t_points = seq(0, 10, by = 0.1), d = 1, alpha = 0.05)

}
